// src/types/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      user_profiles: {
        Row: {
          calendar_token: string | null;
          created_at: string;
          currency: string | null;
          default_wage: number | null;
          first_name: string | null;
          id: number;
          last_name: string | null;
          monthly_email: boolean | null;
          onboarding_completed: boolean | null;
          reminders: boolean | null;
          time_format: string | null;
          updated_at: string;
          user_id: string | null;
          weekly_email: boolean | null;
          work_week: string | null;
        };
        Insert: {
          calendar_token?: string | null;
          created_at?: string;
          currency?: string | null;
          default_wage?: number | null;
          first_name?: string | null;
          id?: number;
          last_name?: string | null;
          monthly_email?: boolean | null;
          onboarding_completed?: boolean | null;
          reminders?: boolean | null;
          time_format?: string | null;
          updated_at?: string;
          user_id?: string | null;
          weekly_email?: boolean | null;
          work_week?: string | null;
        };
        Update: {
          calendar_token?: string | null;
          created_at?: string;
          currency?: string | null;
          default_wage?: number | null;
          first_name?: string | null;
          id?: number;
          last_name?: string | null;
          monthly_email?: boolean | null;
          onboarding_completed?: boolean | null;
          reminders?: boolean | null;
          time_format?: string | null;
          updated_at?: string;
          user_id?: string | null;
          weekly_email?: boolean | null;
          work_week?: string | null;
        };
        Relationships: [];
      };
      work_logs: {
        Row: {
          created_at: string;
          custom_rate: number | null;
          date: string | null;
          default_rate: boolean | null;
          end_time: string | null;
          id: number;
          notes: string | null;
          organisation_id: string | null;
          paid: boolean;
          start_time: string | null;
          updated_at: string;
          user_id: string | null;
        };
        Insert: {
          created_at?: string;
          custom_rate?: number | null;
          date?: string | null;
          default_rate?: boolean | null;
          end_time?: string | null;
          id?: number;
          notes?: string | null;
          organisation_id?: string | null;
          paid?: boolean;
          start_time?: string | null;
          updated_at?: string;
          user_id?: string | null;
        };
        Update: {
          created_at?: string;
          custom_rate?: number | null;
          date?: string | null;
          default_rate?: boolean | null;
          end_time?: string | null;
          id?: number;
          notes?: string | null;
          organisation_id?: string | null;
          paid?: boolean;
          start_time?: string | null;
          updated_at?: string;
          user_id?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "work_logs_organisation_id_fkey";
            columns: ["organisation_id"];
            isOneToOne: false;
            referencedRelation: "organisations";
            referencedColumns: ["id"];
          },
        ];
      };
    };
// ...
  };
};